requirements.txt
annotated-types==0.6.0
anyio==4.3.0
bertopic==0.16.2
blinker==1.8.2
blis==0.7.11
catalogue==2.0.10
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
confection==0.1.4
cymem==2.0.8
Cython==0.29.37
distro==1.9.0
filelock==3.14.0
Flask==3.0.3
fsspec==2024.5.0
greenlet==3.0.3
h11==0.14.0
hdbscan==0.8.33
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
langcodes==3.4.0
language_data==1.2.0
llvmlite==0.42.0
marisa-trie==1.1.1
MarkupSafe==2.1.5
mpmath==1.3.0
murmurhash==1.0.10
networkx==3.3
numba==0.59.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
openai==1.30.1
packaging==24.0
pandas==2.2.2
pillow==10.3.0
plotly==5.22.0
preshed==3.0.9
psycopg2-binary==2.9.9
pydantic==2.7.1
pydantic_core==2.18.2
pynndescent==0.5.12
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.31.0
safetensors==0.4.3
scikit-learn==1.4.2
scipy==1.13.0
sentence-transformers==2.7.0
six==1.16.0
smart-open==6.4.0
sniffio==1.3.1
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.30
srsly==2.4.8
sympy==1.12
tenacity==8.3.0
thinc==8.2.3
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.3.0
tqdm==4.66.4
transformers==4.41.0
triton==2.3.0
typer==0.9.4
typing_extensions==4.11.0
tzdata==2024.1
umap-learn==0.5.6
urllib3==2.2.1
waitress==3.0.0
wasabi==1.1.2
weasel==0.3.4
Werkzeug==3.0.3
